# specification-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Store API - OpenAPI 3.0
  description: |-
    This is a Store API Server
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://store-api.swagger.io/api/v3
...
paths:
...
  /batch:
    put:
      tags:
        - batch
      summary: Update an existing Batch
      description: Update an existing Batch by Id
      operationId: updateBatch
      requestBody:
        description: Update an existent Batch in the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/BatchDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BatchDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/BatchResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Batch not found
        '405':
          description: Validation exception
      security:
        - bearerAuth: []
    post:
      tags:
        - batch
      summary: Add a new Batch to the app
      description: Add a new Batch to the app
      operationId: addBatch
      requestBody:
        description: Create a new Batch to the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/BatchDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BatchDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/BatchResponseDto'
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
  /batch/findByStatus:
    get:
      tags:
        - batch
      summary: Finds Batch by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findBatchByStatus
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: false
          explode: true
          schema:
            type: string
            default: ACTIVE
            enum:
              - ACTIVE
              - DEACTIVE
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchResponseDto'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchResponseDto'
        '400':
          description: Invalid status value
      security:
        - bearerAuth: []
  /batch/{batchId}:
    get:
      tags:
        - batch
      summary: Find Batch by ID
      description: Returns a single Batch
      operationId: getBatchById
      parameters:
        - name: batchId
          in: path
          description: ID of Batch to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/BatchResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Batch not found
      security:
        - bearerAuth: []
    post:
      tags:
        - batch
      summary: Updates a Batch in the app with form data
      description: ''
      operationId: updateBatchWithForm
      parameters:
        - name: batchId
          in: path
          description: ID of Batch to return
          required: true
          schema:
            type: integer
            format: int64
        - name: batchName
          in: query
          description: Name of Batch that needs to be updated
          schema:
            type: string
        - name: buyingPrice
          in: query
          description: Buying Price of Batch that needs to be updated
          schema:
            type: number
            format: double
        - name: sellingPrice
          in: query
          description: Buying Price of Batch that needs to be updated
          schema:
            type: number
            format: double
        - name: status
          in: query
          description: Status of Batch that needs to be updated
          schema:
            type: string
      responses:
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
    delete:
      tags:
        - batch
      summary: Deletes a Batch
      description: delete a Batch
      operationId: deleteBatch
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: batchId
          in: path
          description: Batch id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid Batch value
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    SkuDto:
      required:
        - skuName
      type: object
      properties:
        skuId:
          type: integer
          format: int64
          example: 10
        skuName:
          type: string
          example: 2L, 4L
        photoUrl:
          type: string
        productId:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: SKU status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: SkuDto
...
    BatchDto:
      required:
        - batchName
        - sellingPrice
        - skuId
      type: object
      properties:
        batchId:
          type: integer
          format: int64
          example: 10
        batchName:
          type: string
          example: 1st batch, 2nd batch
        buyingPrice:
          type: number
          format: double
          example: 100.00
        sellingPrice:
          type: number
          format: double
          example: 100.00
        photoUrl:
          type: string
        skuId:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: Batch status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: batch
    BatchResponseDto:
      required:
        - batchName
        - sellingPrice
        - skuId
      type: object
      properties:
        batchId:
          type: integer
          format: int64
          example: 10
        batchName:
          type: string
          example: 1st batch, 2nd batch
        buyingPrice:
          type: number
          format: double
          example: 100.00
        sellingPrice:
          type: number
          format: double
          example: 100.00
        photoUrl:
          type: string
        sku:
          $ref: '#/components/schemas/SkuDto'
        status:
          type: string
          description: Batch status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: BatchResponseDto
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT